Split sidebar provider kit models by type
