+++ Packages/CmuxSidebarProviderKit/Sources/CmuxSidebarProviderKit/Rows/CmuxSidebarProviderRowAccessoryKind.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+/// Accessory kind shown on a provider row.
+public enum CmuxSidebarProviderRowAccessoryKind: String, Codable, Equatable, Sendable {
+    /// Opens CMUX's workspace inspector affordance.
+    case workspaceInspector
+}
