+++ Packages/CmuxSidebarProviderKit/Sources/CmuxSidebarProviderKit/Text/CmuxSidebarProviderRelativeDateStyle.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+/// Relative-date formatting style for provider text.
+public enum CmuxSidebarProviderRelativeDateStyle: String, Codable, Equatable, Sendable {
+    /// Compact elapsed-time style for dense sidebar rows.
+    case compact
+}
